chore: move utils file from lib to libs
