+++ src/blocks/SolarBlock/SolarBlock.tsx
@@ -8,7 +8,7 @@ import { TrustBlock } from '@/blocks/TrustBlock/TrustBlock';
 import { Section } from '@/components/Section/Section';
 import { Button } from '@/components/ui/button';
 import { useVideoPlayer } from '@/hooks/useVideoPlayer';
-import { cn } from '@/lib/utils';
+import { cn } from '@/libs/utils';
 import { useProducts } from '@/services/api/dto/Product/Product.query';
 import { VIDEOS } from '@/services/api/uploadThingFiles';
