+++ src/blocks/TrustBlock/TrustBlock.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/lib/utils';
+import { cn } from '@/libs/utils';
 
 type TrustBlockProps = {
   title: string;
